requirements.txt
# Core deep learning
torch>=2.2.0
numpy>=1.24.0

# Dataset access
tslearn>=0.6.3

# Foundation model
chronos-forecasting>=1.3.0      # pip install chronos-forecasting

# Classical ML probes
scikit-learn>=1.4.0

# Visualisation
matplotlib>=3.8.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
